# backend/internal/httpapi/panelsettings/docs/swagger.yaml
basePath: /api
definitions:
...
  shared.ErrorResponse:
    properties:
      error:
        example: invalid input
        type: string
      message:
        example: invalid input
        type: string
    type: object
...
info:
  contact: {}
  description: Exodus dashboard and node-management API.
  license:
    name: AGPL-3.0
  title: Exodus API
  version: 3.0.0
paths:
...
  /exodus-settings:
    get:
      consumes:
      - application/json
      description: Get or update Exodus panel branding, passkey, oauth2, and password
        policies
      parameters:
      - description: Settings payload to update
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage Exodus panel settings
      tags:
      - Exodus Settings Controller
    patch:
      consumes:
      - application/json
      description: Get or update Exodus panel branding, passkey, oauth2, and password
        policies
      parameters:
      - description: Settings payload to update
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage Exodus panel settings
      tags:
      - Exodus Settings Controller
...
  /tokens:
    get:
      consumes:
      - application/json
      description: List or create (201) scoped panel API tokens
      parameters:
      - description: API token creation parameters
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage API tokens
      tags:
      - API Tokens Controller
    post:
      consumes:
      - application/json
      description: List or create (201) scoped panel API tokens
      parameters:
      - description: API token creation parameters
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage API tokens
      tags:
      - API Tokens Controller
  /tokens/{uuid}:
    delete:
      description: Revoke API token by token UUID
      parameters:
      - description: Token UUID
        format: uuid
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete API token
      tags:
      - API Tokens Controller
  /tokens/ott:
    post:
      description: Issue short-lived token for internal utilities and redirects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Issue One-Time Token (OTT)
      tags:
      - API Tokens Controller
  /tokens/scopes:
    get:
      description: Get list of all supported permission scopes for panel tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get available API token scopes
      tags:
      - API Tokens Controller
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"